minor cast fixes to prevent 64bit warnings

+++ unix/tkUnixScale.c
@@ -10,7 +10,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tkUnixScale.c,v 1.7 2001/07/03 06:08:33 hobbs Exp $
+ * RCS: @(#) $Id: tkUnixScale.c,v 1.8 2001/09/21 21:34:10 hobbs Exp $
  */
 
 #include "tkScale.h"
@@ -271,7 +271,7 @@ DisplayVerticalValue(scalePtr, drawable, value, rightEdge)
     Tk_GetFontMetrics(scalePtr->tkfont, &fm);
     y = TkScaleValueToPixel(scalePtr, value) + fm.ascent/2;
     sprintf(valueString, scalePtr->format, value);
-    length = strlen(valueString);
+    length = (int) strlen(valueString);
     width = Tk_TextWidth(scalePtr->tkfont, valueString, length);
 
     /*
@@ -488,7 +488,7 @@ DisplayHorizontalValue(scalePtr, drawable, value, top)
     Tk_GetFontMetrics(scalePtr->tkfont, &fm);
     y = top + fm.ascent;
     sprintf(valueString, scalePtr->format, value);
-    length = strlen(valueString);
+    length = (int) strlen(valueString);
     width = Tk_TextWidth(scalePtr->tkfont, valueString, length);
 
     /*
